+++ Event/ExtbasePersistenceClassesEvent.php
@@ -25,10 +25,7 @@ namespace Bartacus\Bundle\BartacusBundle\Event;
 
 use Symfony\Contracts\EventDispatcher\Event;
 
-/**
- * @deprecated since 3.0.3, will be removed in 3.1.0
- */
-final class RequestExtbasePersistenceClassesEvent extends Event
+final class ExtbasePersistenceClassesEvent extends Event
 {
     /**
      * @var array