+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Initialize TSFE for symfony routes
+- Naming of extbasePersistenceClasses event and loader method
+### Added
+- `TypoScriptFrontendInitialization` to services
+### Deprecated
+- Classes: `RequestExtbasePersistenceClassesEvent`
+- Methods: `ConfigLoader::loadFromRequestExtbasePersistenceClasses`
+
 ## [3.0.3] - 2020-05-14
 ### Removed
 - Override of TYPO3 error handler